Data type changed

+++ src/Test_Git.java
@@ -1,7 +1,7 @@
 
     public class Test_Git {
-        static double w = 6;
-        static double l = 6;
+        static int w = 6;
+        static int l = 6;
         static double a = w * l;
 
         public static void main(String[] args) {
